+++ app/commands/characters_context/dnd5/make_short_rest_command.rb
@@ -45,6 +45,7 @@ module CharactersContext
       def do_persist(input)
         input[:character].feats.joins(:feat).where.not(feats: { origin: 6 }).where(limit_refresh: 0).update_all(used_count: 0)
         update_character(input)
+        refresh_resources(input)
 
         { result: :ok }
       end
@@ -55,6 +56,24 @@ module CharactersContext
         input[:character].data.spent_hit_dice.merge!(input[:spent_hit_dice])
         input[:character].data.health['current'] = input[:health] if input[:health]
         input[:character].save
+      end
+
+      def refresh_resources(input) # rubocop: disable Metrics/AbcSize
+        input[:character].resources.includes(:custom_resource).find_each do |resource|
+          max_value = resource.custom_resource.max_value
+          reset_direction = resource.custom_resource.reset_direction
+          change = resource.custom_resource.resets['short']
+
+          value =
+            case change
+            when -1 then reset_direction.zero? ? 0 : max_value
+            when 0 then resource.value
+            else
+              reset_direction.zero? ? [resource.value - change.abs, 0].max : [resource.value + change.abs, max_value].min
+            end
+
+          resource.update(value: value)
+        end
       end
     end
   end